Move BatchRequest tests into batch_request.py
