+++ tests/checkpoint/test_checkpoint_with_fluent_datasources.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import copy
 import logging
 import pickle
